fix vllm ci
use post1
