+++ pyproject.toml
@@ -29,7 +29,7 @@ sacrebleu = {extras = ["ja"], version = "^2.4.1"}
 jiwer = "^3.0.4"
 openai = "^1.52.2"
 google-api-python-client = "^2.131.0"
-vllm = {version = "^0.6.4", optional = true }
+vllm = {version = "^0.6.4.post1", optional = true }
 loguru = "^0.7.2"
 wandb = {version = "^0.17.2", optional = true}
 pyarrow = "16.1.0"  # set the version because we get "Unable to find installation candidates" with 17.0.0
