+++ tests/conftest.py
@@ -1,0 +1,11 @@
+import pytest
+
+
+def is_vllm_enabled() -> bool:
+    try:
+        import torch
+        import vllm  # noqa: F401
+
+        return torch.cuda.device_count() > 0
+    except ImportError:
+        return False
